server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.7
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
